+++ tests/charts/bootstrap.sh
@@ -13,9 +13,15 @@ python -m pip install pyyaml==6.0.1 \
 
 cd ..
 helm template dummy --values tests/charts/templates/render/dummy.yaml \
-  charts/selenium-grid > ./tests/tests/output_deployment.yaml
+  --set-file 'nodeConfigMap.extraScripts.nodePreStop\.sh=tests/charts/templates/render/dummy_external.sh' \
+  --set-file 'recorderConfigMap.extraScripts.video\.sh=tests/charts/templates/render/dummy_external.sh' \
+  --set-file 'recorderConfigMap.extraScripts.graphQLRecordVideo\.sh=tests/charts/templates/render/dummy_external.sh' \
+  --set-file 'recorderConfigMap.extraScripts.newInsertScript\.sh=tests/charts/templates/render/dummy_external.sh' \
+  --set-file 'uploaderConfigMap.extraScripts.entry_point\.sh=tests/charts/templates/render/dummy_external.sh' \
+  --set-file 'uploaderConfigMap.secretFiles.config\.conf=tests/charts/templates/render/dummy_external.sh' \
+  charts/selenium-grid > ./tests/tests/dummy_template_manifests.yaml
 
-python tests/charts/templates/test.py "./tests/tests/output_deployment.yaml"
+python tests/charts/templates/test.py "./tests/tests/dummy_template_manifests.yaml"
 ret_code=$?
 
 if [ "${CI:-false}" = "false" ]; then