+++ tests/charts/make/chart_cluster_setup.sh
@@ -38,15 +38,8 @@ kind create cluster --wait ${WAIT_TIMEOUT} --name ${CLUSTER_NAME} --config tests
 echo "Install KEDA core on kind kubernetes cluster"
 kubectl apply --server-side -f https://github.com/kedacore/keda/releases/download/v2.12.1/keda-2.12.1-core.yaml
 
-echo "Install ingress-nginx on kind kubernetes cluster"
-kubectl apply -n ${INGRESS_NAMESPACE} -f https://raw.githubusercontent.com/kubernetes/ingress-nginx/main/deploy/static/provider/kind/deploy.yaml
-kubectl wait --namespace ${INGRESS_NAMESPACE} \
-  --for=condition=ready pod \
-  --selector=app.kubernetes.io/component=controller \
-  --timeout=${WAIT_TIMEOUT}
-
 echo "Load built local Docker Images into Kind Cluster"
-image_list=$(docker images --format "{{.Repository}}:{{.Tag}}" | grep ${NAMESPACE} | grep ${VERSION})
+image_list=$(docker images --format "{{.Repository}}:{{.Tag}}" | grep ${NAMESPACE} | grep ${BUILD_DATE:-$VERSION})
 for image in $image_list; do
     kind load docker-image --name ${CLUSTER_NAME} "$image"
 done