+++ tests/charts/templates/render/dummy_external.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+echo "Your custom script patch to chart rendering"